requirements.txt
Flask==2.0.3
mysql-connector-python==8.0.28
flask-wtf==1.2.1
wtforms==3.1.2
email-validator==2.2.0
datetime==5.5